dsv41: gemv cp.async weights-first (P3 from gemm-prologue-overlap) - cp.async.cg.shared.global 16B loads for weight rows, issued BEFORE the activation staging prologue (weights don't depend on activation). 5-deep pipeline with independent staging buffer. The activation staging/LUT/a32 materialization overlaps with weight prefetch. Expected: prologue 7.3->5us x 246 = -0.57ms. Gate: DSV41_GEMV_CPASYNC (default ON, =0 reverts). Includes unsigned cast fix for pf_row ternary.

+++ kernels/cuda/dsv41_kernels.cu
@@ -3249,7 +3249,10 @@ gemm_fp8_gemv_kernel(__grid_constant__ const GemvCore gc, __grid_constant__ cons
     // same slot) -- the only difference is WHEN the instruction is issued. It is
     // issued after cudaGridDependencySynchronize(), like every other read of a
     // producer's output, because `w` can be written by the previous node.
-    int pf_row = ((cpasync != 0) && (vec >= 3)) ? (blockIdx.x * nwarps + warp) : -1;
+    // NOTE the explicit (int) cast: `blockIdx.x * nwarps + warp` is UNSIGNED
+    // (blockIdx.x is uint3), so the bare ternary would promote -1 to 0xFFFFFFFF
+    // and leave pf_row == -1 holding on an implementation detail.
+    int pf_row = ((cpasync != 0) && (vec >= 3)) ? ((int)blockIdx.x * nwarps + warp) : -1;
     if (pf_row >= n) pf_row = -1;   // the last block's tail warps have no row
     if (pf_row >= 0) {
         // Family/pointer select copied from the row loop verbatim: a family-2
